ci: enable additional golangci-lint linters and fix violations
Enable bodyclose, containedctx, contextcheck, errorlint, nilerr, noctx,
testifylint, unconvert, and wrapcheck linters in golangci-lint config.

Fix all violations: use errors.Is/errors.As instead of direct comparison
and type assertions, use context-aware net.Listen in production code,
thread ctx through sameSplitOutputPath and openUniqOutput, remove
unnecessary type conversions, and suppress intentional patterns with
targeted nolint directives.

+++ internal/builtins/permission_helpers.go
@@ -109,8 +109,8 @@ func seedPermissionIdentityDBFromEnv(db *permissionIdentityDB, inv *Invocation) 
 	if db == nil || inv == nil {
 		return
 	}
-	uid := uint32(idUintEnv(inv.Env, "UID", idDefaultUID))
-	gid := uint32(idUintEnv(inv.Env, "GID", idDefaultGID))
+	uid := idUintEnv(inv.Env, "UID", idDefaultUID)
+	gid := idUintEnv(inv.Env, "GID", idDefaultGID)
 	user := strings.TrimSpace(inv.Env["USER"])
 	if user == "" {
 		user = idDefaultUserName
